Move env to utils
